Refactor: Integrate rewarded ads into interpretation component
