+++ app/[locale]/reading/page.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next"
 import ReadingType from "@/components/reading/reading-type"
 import CardSelection from "@/components/reading/card-selection"
-import AdViewing from "@/components/reading/ad-viewing"
 import Interpretation from "@/components/reading/interpretation"
 import ReadingGuard from "@/components/reading/reading-guard"
 import { getTranslations } from "next-intl/server"
@@ -57,7 +56,6 @@ export default async function ReadingPage() {
         <ReadingGuard>
             <ReadingType readingConfig={readingConfig} />
             <CardSelection readingConfig={readingConfig} />
-            <AdViewing />
             <Interpretation />
         </ReadingGuard>
     )
